Update the Player controller & added wave/rest timer
In this commit, I added updated the controller by making the camera at the right angle, I made it orthogonal but it can also be changed to perspective in inspector of Camera3D. I also implemented the door spawning and wave timer system. A skip button is also there to skip the rest time and move to wave time instantly.